Changed int to bigint

+++ MySql/mydb.sql
@@ -19,14 +19,14 @@ DROP TABLE IF EXISTS `groups_tests`;
 
 -- creating tables
 CREATE TABLE `users` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
   `user_name` varchar(64) NOT NULL,
   `user_password` varchar(64) NOT NULL
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `files` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
-  `id_user` int(11) NOT NULL,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
+  `id_user` bigint(11) NOT NULL,
   `hash` varchar(64) NOT NULL,
   `name` varchar(256) DEFAULT NULL,
   `file_system_path` varchar(1024),
@@ -34,9 +34,9 @@ CREATE TABLE `files` (
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `tests` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
-  `id_file` int(11) NOT NULL,
-  `id_user` int(11) NOT NULL,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
+  `id_file` bigint(11) NOT NULL,
+  `id_user` bigint(11) NOT NULL,
   `time_of_add` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `test_table` varchar(16) NOT NULL, /*name of table with test parameters*/
   `loaded` tinyint(1) NOT NULL DEFAULT 0,  /*if test was loaded by external programm*/
@@ -51,40 +51,40 @@ CREATE TABLE `tests` (
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `currently_running` (
-  `id_test` int(11) PRIMARY KEY,
+  `id_test` bigint(11) PRIMARY KEY,
   CONSTRAINT currently_running_fk_test_id FOREIGN KEY (`id_test`) REFERENCES `tests`(`id`)
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `nist_tests` (
-  `id_test` int(11) PRIMARY KEY,
-  `length` int(11) NOT NULL,
+  `id_test` bigint(11) PRIMARY KEY,
+  `length` bigint(11) NOT NULL,
   `test_number` int(5) NOT NULL,
-  `streams` int(11),
-  `special_parameter` int(11),
+  `streams` bigint(11),
+  `special_parameter` bigint(11),
   CONSTRAINT nist_tests_fk_test_id FOREIGN KEY (`id_test`) REFERENCES `tests`(`id`)
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `results` (
-  `id_test` int(11) PRIMARY KEY,
+  `id_test` bigint(11) PRIMARY KEY,
   `directory` varchar(1024) NOT NULL,
   CONSTRAINT results_fk_test_id FOREIGN KEY (`id_test`) REFERENCES `tests`(`id`)
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `groups` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
-  `id_user` int(11),
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
+  `id_user` bigint(11),
   `time_of_add` timestamp DEFAULT CURRENT_TIMESTAMP,
   CONSTRAINT groups_fk_id_user FOREIGN KEY (`id_user`) REFERENCES `users`(`id`)
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `groups_tests` (
-  `id` int(11),
-  `id_test` int(11),
+  `id` bigint(11),
+  `id_test` bigint(11),
   CONSTRAINT groups_tests_fk_group_id FOREIGN KEY (`id`) REFERENCES `groups`(`id`),
   CONSTRAINT groups_tests_fk_test_id FOREIGN KEY (`id_test`) REFERENCES `tests`(`id`)
 )DEFAULT CHARSET=utf8;
 
 CREATE TABLE `pid_table` (
-  `id` int(11) PRIMARY KEY,
+  `id` bigint(11) PRIMARY KEY,
   `pid` int(11) NOT NULL
 )DEFAULT CHARSET=utf8;
